% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPs2EnrichR.R
\name{DEPs2EncirhR}
\alias{DEPs2EncirhR}
\title{DEPs to EnrichR}
\usage{
DEPs2EncirhR(DEPs, GOdb)
}
\arguments{
\item{DEPs}{as Data Frame that have "GeneName" Column in dataframe, essential}

\item{GOdb}{Define by enrichR pacakge, find from "https://maayanlab.cloud/Enrichr/#libraries"}
}
\value{
return Gene ontology results,  you can any database in EnrichR
}
\description{
DEPs to EnrichR
}
\examples{
DEPs2EncirhR(DEPs, GOdb)
}
