% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichR2GOplot-package.R
\docType{package}
\name{EnrichR2GOplot-package}
\alias{EnrichR2GOplot}
\alias{EnrichR2GOplot-package}
\title{EnrichR2GOplot: Gene ontology analysis with EnrichR, Convert Z-score with GOplot pkg}
\description{
Gene Ontology analysis pipeline with EnrichR and GOplots.
}
\author{
\strong{Maintainer}: Seongmin Cheon \email{s.cheon1995@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
