% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichR2GOplot.R
\name{EnrichR2GOplot}
\alias{EnrichR2GOplot}
\title{EnrichR2GOplot}
\usage{
EnrichR2GOplot(DEPs, GOdb, cutoff)
}
\arguments{
\item{DEPs}{dataframe DEPs have two columns that "GeneName" and "logFC"}

\item{GOdb}{Working on default GOdb setting, GOdb <- c("GO_Biological_Process_2023", "GO_Cellular_Component_2023", "GO_Molecular_Function_2023","KEGG_2021_Human")}

\item{cutoff}{minimum significant adj.p-value for EnrichR  enrichment analysis}
}
\value{
Return Results of list, Gene Ontology, KEGG, and GOplot input (circ)
}
\description{
EnrichR2GOplot
}
\examples{
GOdb <- c("GO_Biological_Process_2023", "GO_Cellular_Component_2023", "GO_Molecular_Function_2023","KEGG_2021_Human")
colnames(DEPs) <- c("GeneName","logFC")
EnrichR2GOplot(DEPs, GOdb, 0.05)



}
