% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table_EnrichR2GOplot.R
\name{export_table_EnrichR2GOplot}
\alias{export_table_EnrichR2GOplot}
\title{Export table from EnrichR2GOplot results}
\usage{
export_table_EnrichR2GOplot(export_from_EnrichR2GOplot, DEPs, cutoff)
}
\arguments{
\item{export_from_EnrichR2GOplot}{return from EnrichR2GOplot()}

\item{DEPs}{dataframe DEPs have two columns that "GeneName" and "logFC"}

\item{cutoff}{minimum significant adj.p-value for EnrichR  enrichment analysis}
}
\value{
export table format GO result and KEGG results
}
\description{
Export table from EnrichR2GOplot results
}
\examples{

result <- EnrichR2GOplot(DEPs,GOdb, 0.05)

export <- export_table_EnrichR2GOplot(result, DEPs, 0.05)

}
