% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isJobReady.R
\name{isJobReady}
\alias{isJobReady}
\title{jobID for Uniprot accession to gene name}
\usage{
isJobReady(jobId)
}
\arguments{
\item{jobId}{jobID , function automatically call out by other functions}
}
\value{
return jobID to function UniprotAccession2GeneID
}
\description{
jobID for Uniprot accession to gene name
}
\examples{
isJobReady(jobID)
}
