% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotAccession2GeneName.R
\name{UniprotAccession2GeneName}
\alias{UniprotAccession2GeneName}
\title{UniprotAccession2GeneName}
\usage{
UniprotAccession2GeneName(Accession)
}
\arguments{
\item{Accession}{Uniprot Accession ID,  support to isoforms format Oxxxx-1 , Import by Vector format ,  Verification required for legacy Accession number such as A00XXXX}
}
\value{
Return dataframe that have Three columns "Input accession" ,
}
\description{
UniprotAccession2GeneName
}
\examples{
Accession <- c("O95905-1","P04439-2","P08246")
UniprotAccession2GeneName(Accession)

}
